fix(lint): split teacher review detail layout components
